+++ NEWS.md
@@ -1,6 +1,6 @@
 # parsnip (development version)
 
- * An RStudio add-in is availble that makes writing multiple `parsnip` model specifications to the source window. It can be accessed via the IDE addin menus or by calling `write_parsnip_specs()`.
+ * An RStudio add-in is availble that makes writing multiple `parsnip` model specifications to the source window. It can be accessed via the IDE addin menus or by calling `parsnip_addin()`.
  
 # parsnip 0.1.4